pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
authlib = "^1.3.0"
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
fastapi-users = "^12.1.2"
httpx-oauth = "^0.10.2"
fastapi-users-db-sqlalchemy = "^6.0.1"
python-dotenv = "^1.0.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
alembic = "^1.11.1"
asyncpg = {version = "^0.28.0", extras = ["sa"]}
aiofiles = "^23.1.0"
httptools = "^0.6.0"
prometheus-client = "^0.17.0"
prometheus-fastapi-instrumentator = "6.0.0"
fastapi-pagination = "^0.12.24"
google-auth = "^2.29.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
itsdangerous = "^2.2.0"
catilo = "^0.2.6"
cachetools = "^5.3.3"
fastapi-filter = "^2.0.0"
pandas = "^2.2.2"
bs4 = "^0.0.2"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "4.0.0"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^24.4.2"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]


[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["backend",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "BACKEND_ENVIRONMENT=pytest",
    "BACKEND_DB_BASE=backend_test",
]

[fastapi-template.options]
project_name = "backend"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "github"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "True"
kube_name = "backend"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "None"
add_dummy = "None"
orm = "sqlalchemy"
self_hosted_swagger = "True"
prometheus_enabled = "True"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "None"
add_users = "True"
cookie_auth = "True"
jwt_auth = "True"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"